Fix casting warnings on 32-bit builds.
AlignTo is now safe to use with 32-bit multiples and fixed other number type mismatches.

+++ src/gpgmm/common/SlabMemoryAllocator.h
@@ -130,7 +130,7 @@ namespace gpgmm {
             explicit SlabAllocatorCacheEntry(uint64_t blockSize) : mBlockSize(blockSize) {
             }
 
-            size_t GetKey() const {
+            uint64_t GetKey() const {
                 return mBlockSize;
             }
